docker-compose/monitor/requirements.txt
Deprecated==1.2.13
googleapis-common-protos==1.56.2
grpcio==1.50.0
opentelemetry-api==1.13.0
opentelemetry-exporter-jaeger==1.13.0
opentelemetry-exporter-jaeger-proto-grpc==1.13.0
opentelemetry-exporter-jaeger-thrift==1.13.0
opentelemetry-sdk==1.13.0
opentelemetry-semantic-conventions==0.34b0
protobuf==3.20.3
six==1.16.0
thrift==0.16.0
typing_extensions==4.4.0
wrapt==1.14.1